Add main watch lists page
