+++ app/assets/stylesheets/application.sass.scss
@@ -69,3 +69,4 @@ $govuk-images-path: '/govuk-frontend/dist/govuk/assets/images/';
 @import 'src/faq';
 @import 'src/chapter_selection';
 @import 'src/mycommodities';
+@import 'src/app-card';
