+++ app/assets/stylesheets/src/_app-card.scss
@@ -1,0 +1,13 @@
+.app-card {
+  background-color: #f8f8f8;
+  display: block;
+  padding: 20px;
+  margin-bottom: 20px;
+  @include govuk-media-query(desktop) {
+    min-height: 160px;
+  }
+}
+
+.app-card p:last-child {
+  margin-bottom: 0;
+}
