+++ app/views/myott/preferences/edit.html.erb
@@ -1,7 +1,7 @@
 <% myott_page_title "Select chapters", error: flash.now[:alert] %>
 
 <% content_for :back_link do %>
-  <% back_link session[:all_tariff_updates] ? myott_check_your_answers_path : myott_path %>
+  <% back_link session[:all_tariff_updates] ? myott_stop_press_check_your_answers_path : myott_path %>
 <% end %>
 
 <div class="govuk-grid-row">
@@ -11,7 +11,7 @@
       <div class="govuk-body myott-subscriptions">
         <p>Products are broken down into sections and chapters. Sections broadly group products, while chapters get more specific. You can choose to be updated about as many chapters as you like.</p>
         <%= render partial: 'checkbox_toggles' %>
-        <%= form_with url: myott_preferences_path, method: :patch, local: true do %>
+        <%= form_with url: myott_stop_press_preferences_path, method: :patch, local: true do %>
           <div class="govuk-accordion" data-module="govuk-accordion" id="section-chapter-accordion">
             <% @all_sections_chapters.each do |section, chapters| %>
               <div class="govuk-accordion__section <%= section.resource_id == "1" ? 'govuk-accordion__section--expanded' : ''%>" id="accordion-section-<%=section.resource_id%>">
