+++ app/views/myott/preferences/new.html.erb
@@ -8,7 +8,7 @@
       <p class="govuk-inset-text">You can change what tariff chapters you receive updates about at any time
         by following the link in the tariff update emails.</p>
 
-      <%= form_with url: myott_preferences_path, method: :post, local: true do %>
+      <%= form_with url: myott_stop_press_preferences_path, method: :post, local: true do %>
         <fieldset class="govuk-fieldset" aria-describedby="select-hint">
           <legend class="govuk-fieldset__legend govuk-fieldset__legend--m">
             <h2 class="govuk-fieldset__heading govuk-heading-m">What tariff updates would you like to receive?</h2>
