+++ app/views/myott/stop_presses/_details_selected_sections_chapters.html.erb
