+++ app/views/myott/stop_presses/check_your_answers.html.erb
@@ -1,7 +1,7 @@
 <% myott_page_title "Review selection", error: flash.now[:alert] %>
 
 <% content_for :back_link do %>
-  <% back_link session[:all_tariff_updates] ? new_myott_preferences_path : edit_myott_preferences_path %>
+  <% back_link session[:all_tariff_updates] ? new_myott_stop_press_preferences_path : edit_myott_stop_press_preferences_path %>
 <% end %>
 
 <div class="govuk-body myott-subscriptions">
@@ -16,9 +16,9 @@
       </p>
 
       <div data-controller="chapter-selection">
-        <%= form_with url: myott_subscribe_path, method: :post, local: true do %>
+        <%= form_with url: subscribe_myott_stop_press_path, method: :post, local: true do %>
           <%= render 'details_selected_sections_chapters', selected_sections_chapters: @selected_sections_chapters %>
-          <%= link_to "Edit your chapter preferences", edit_myott_preferences_path, class: "link_button", style: "font-size: 19px;" %>
+          <%= link_to "Edit your chapter preferences", edit_myott_stop_press_preferences_path, class: "link_button", style: "font-size: 19px;" %>
           <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible">
 
           <h2 class="govuk-heading-m">Confirm your selection</h2>
