+++ app/views/myott/stop_presses/confirmation.html.erb
