+++ app/views/myott/stop_presses/show.html.erb
@@ -10,7 +10,7 @@
 
     <%= render 'details_selected_sections_chapters', selected_sections_chapters: @selected_sections_chapters %>
 
-    <p class="govuk-body"><a href=<%= edit_myott_preferences_path %>>Edit your chapter preferences</a></p>
+    <p class="govuk-body"><a href=<%= edit_myott_stop_press_preferences_path %>>Edit your chapter preferences</a></p>
 
     <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible">
