+++ app/views/myott/subscriptions/index.html.erb
@@ -1,0 +1,37 @@
+<% myott_page_title "Your tariff watch lists" %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-full">
+    <h1 class="govuk-heading-l">Your tariff watch lists</h1>
+    <p class="govuk-body"><strong>Email address:</strong> <%= @current_user.email %></p>
+  </div>
+
+
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-one-half">
+      <div class="app-card">
+        <h2 class="govuk-heading-m app-card__heading">Commodity watch list</h2>
+        <% if @my_commodities.present? %>
+          <%= link_to 'View commodity watch list', myott_mycommodities_path %>
+        <% else %>
+          <p class="govuk-body"><strong>No commodities selected</strong></p>
+          <p class="govuk-body">You will receive updates on specific commodities</p>
+          <%= link_to 'Create a commodity watch list', myott_mycommodities_path %>
+        <% end %>
+      </div>
+    </div>
+
+    <div class="govuk-grid-column-one-half">
+      <div class="app-card">
+        <h2 class="govuk-heading-m app-card__heading">Stop Press watch list</h2>
+        <% if @stop_press.present? %>
+          <%= link_to 'View Stop Press watch list', myott_stop_press_path %>
+        <% else %>
+          <p class="govuk-body"><strong>No chapters selected</strong></p>
+          <p class="govuk-body">You will receive updates on specific chapters</p>
+          <%= link_to 'Create a Stop Press watch list', myott_stop_press_path %>
+        <% end %>
+      </div>
+    </div>
+  </div>
+</div>
